moved lock history component
