+++ app/src/components/LockedEthList.tsx
@@ -1,0 +1,103 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable react/prop-types */
+import React, { useState, useEffect } from 'react';
+import { getLockEvents, defaultAddress } from '../helpers/lockdrop/EthereumLockdrop';
+//import * as ethAddress from 'ethereum-address';
+import Web3 from 'web3';
+import SectionCard from '../components/SectionCard';
+import { LockEvent } from '../models/LockdropModels';
+import { createStyles, Theme, makeStyles } from '@material-ui/core/styles';
+import List from '@material-ui/core/List';
+import ListItem from '@material-ui/core/ListItem';
+import ListItemText from '@material-ui/core/ListItemText';
+import ListSubheader from '@material-ui/core/ListSubheader';
+import { Divider } from '@material-ui/core';
+
+const useStyles = makeStyles((theme: Theme) =>
+    createStyles({
+        listRoot: {
+            width: '100%',
+            maxWidth: 'auto',
+            backgroundColor: theme.palette.background.paper,
+            position: 'relative',
+            overflow: 'auto',
+            maxHeight: 360,
+        },
+        listSection: {
+            backgroundColor: 'inherit',
+        },
+        ul: {
+            backgroundColor: 'inherit',
+            padding: 0,
+        },
+        lockListPage: {
+            textAlign: 'center',
+        },
+    }),
+);
+
+interface LockHistroyProps {
+    web3: Web3;
+}
+// component that displays the number of tokens and the duration for the lock via Web3
+const LockedEthList: React.FC<LockHistroyProps> = ({ web3 }) => {
+    const classes = useStyles();
+    const [lockEvents, setEvents] = useState<LockEvent[]>([]);
+
+    const updateList = () => {
+        getLockEvents(web3).then(setEvents);
+    };
+
+    useEffect(() => {
+        updateList();
+    }, []);
+
+    return (
+        <>
+            <SectionCard maxWidth="lg">
+                <div className={classes.lockListPage}>
+                    {lockEvents.length > 0 ? (
+                        <>
+                            <h1>Your Lock History</h1>
+                            <List className={classes.listRoot} subheader={<li />}>
+                                <li className={classes.listSection}>
+                                    <ul className={classes.ul}>
+                                        <ListSubheader>You have locked {lockEvents.length} time(s)</ListSubheader>
+                                        <Divider />
+                                        {lockEvents.map(eventItem => (
+                                            <>
+                                                <ListItem key={eventItem.lock}>
+                                                    <ListItemText>
+                                                        <h4>Lock address: {eventItem.lock}</h4>
+                                                        <h5>Locked in block no. {eventItem.blockNo}</h5>
+                                                        <p>
+                                                            Locked {web3.utils.fromWei(eventItem.eth, 'ether')} ETH for{' '}
+                                                            {eventItem.duration} days
+                                                        </p>
+                                                        {eventItem.introducer !== defaultAddress ? (
+                                                            <p>Introducer: {eventItem.introducer}</p>
+                                                        ) : (
+                                                            <p>No introducer</p>
+                                                        )}
+                                                    </ListItemText>
+                                                </ListItem>
+                                                <Divider />
+                                            </>
+                                        ))}
+                                    </ul>
+                                </li>
+                            </List>
+                        </>
+                    ) : (
+                        <>
+                            <h1>No Locks</h1>
+                            <h4>Please lock some ETH!</h4>
+                        </>
+                    )}
+                </div>
+            </SectionCard>
+        </>
+    );
+};
+
+export default LockedEthList;
